# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Jobseek dashboard API
  version: 0.1.0
  description: Versioned private dashboard API. All data routes require authentication.
servers:
  - url: /api/v1
paths:
...
  /rounds/{id}/stop:
    parameters: [{ $ref: '#/components/parameters/RecordId' }]
    post:
      operationId: stopRound
      summary: Owner fences further round work and requests cancellation
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          {
            description: Stopped round,
            content: { application/json: { schema: { $ref: '#/components/schemas/Round' } } },
          }
  /rounds/{id}/resume:
    parameters: [{ $ref: '#/components/parameters/RecordId' }]
    post:
      operationId: resumeRound
      summary: Owner resumes after bounded reconciliation
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          {
            description: Resumed round,
            content: { application/json: { schema: { $ref: '#/components/schemas/Round' } } },
          }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /research/brief:
    get:
      operationId: getSearchBrief
      summary: Read the current search brief with rubric and catalog versions
      description: Versioned owner profile, retained requirements, and reason-catalog version. Side-effect-free.
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          {
            description: Search brief,
            content:
              { application/json: { schema: { $ref: '#/components/schemas/SearchBriefView' } } },
          }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /research/briefs/{version}/catalog:
    parameters:
      - name: version
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
    get:
      operationId: getReasonCatalog
      summary: Read the saved reason catalog for one brief version
      description: Immutable authored positive, negative, and missing-information reason choices. Side-effect-free.
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          {
            description: Reason catalog,
            content:
              { application/json: { schema: { $ref: '#/components/schemas/ReasonCatalogView' } } },
          }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /research/runs/{id}/findings:
    parameters: [{ $ref: '#/components/parameters/RecordId' }]
    get:
      operationId: listRunFindings
      summary: List collected vacancies with saved groups and reasons
      description: One entry per collected vacancy with group, saved reason selections, and evidence links. Side-effect-free.
      security: [{ cookieAuth: [] }]
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
        - name: group
          in: query
          schema:
            type: string
            enum:
              [
                recommended,
                could_be_recommended,
                probably_not_recommended,
                not_recommended,
                unknown,
              ]
      responses:
        '200':
          {
            description: Finding list,
            content: { application/json: { schema: { $ref: '#/components/schemas/FindingList' } } },
          }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /opportunities/{id}/finding:
    parameters: [{ $ref: '#/components/parameters/RecordId' }]
    get:
      operationId: getOpportunityFinding
      summary: Read one collected vacancy's saved classification and reasons
      description: Why-this-job view from saved assessment state. Makes no model call. Side-effect-free.
      security: [{ cookieAuth: [] }]
      responses:
        '200':
          {
            description: Saved finding,
            content:
              { application/json: { schema: { $ref: '#/components/schemas/FindingEntry' } } },
          }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  parameters:
    RecordId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jobseek_session
      description: Owner session. Cookie mutations also require X-CSRF-Token and same-origin Origin.
    bearerAuth:
      type: http
      scheme: bearer
      description: Named agent token with operation-specific scopes.
  responses:
    Unavailable:
      description: Required round execution capability is unavailable
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    ValidationError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Conflict:
      description: Revision conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    RoundAllowance:
      type: object
      required: [requests, items, tools, turns]
      properties:
        requests: { type: integer, format: int64, minimum: 0 }
        items: { type: integer, format: int64, minimum: 0 }
        tools: { type: integer, format: int64, minimum: 0 }
        turns: { type: integer, format: int64, minimum: 0 }
    RoundScope:
      type: object
      required: [inputRefs, resources, operations, delegates]
      properties:
        inputRefs: { type: array, items: { type: string } }
        resources: { type: array, items: { type: string } }
        operations: { type: array, items: { type: string } }
        delegates: { type: array, items: { type: string } }
...
    Round:
      type: object
      required:
        [
          id,
          requestKey,
          intent,
          outcome,
          originalProfileVersion,
          effectiveProfileVersion,
          profileVersion,
          scope,
          state,
          revision,
          generation,
          deadline,
          limits,
          used,
          step,
          cursor,
          unresolved,
          report,
          stopReason,
          deliverableStatus,
          reconciliationRequired,
          createdAt,
          updatedAt,
        ]
      properties:
        id: { type: string }
        requestKey: { type: string }
        intent: { type: string }
        outcome: { type: string }
        originalProfileVersion: { type: integer, format: int64 }
        effectiveProfileVersion: { type: integer, format: int64 }
        profileVersion: { type: integer, format: int64 }
        scope: { $ref: '#/components/schemas/RoundScope' }
        state:
          {
            type: string,
            enum: [queued, running, awaiting_input, stopping, paused, completed, failed],
          }
        revision: { type: integer, format: int64 }
        generation: { type: integer, format: int64 }
        deadline: { type: string, format: date-time }
        limits: { $ref: '#/components/schemas/RoundAllowance' }
        used: { $ref: '#/components/schemas/RoundAllowance' }
        step: { type: string }
        cursor: { type: object, additionalProperties: true }
        unresolved: { type: array, items: {} }
        report: { type: object, additionalProperties: true }
        stopReason: { type: string }
        deliverableStatus: { type: string }
        reconciliationRequired: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
...
    RoleCriterion:
      type: object
      required: [id, label, description, kind, mode]
      properties:
        id: { type: string, minLength: 1, maxLength: 80 }
        label: { type: string, minLength: 1, maxLength: 100 }
        description: { type: string, maxLength: 1000 }
        kind: { type: string, enum: [role, responsibility, technology] }
        mode: { type: string, enum: [require, avoid, prefer] }
    RoleCriterionView:
      allOf:
        - $ref: '#/components/schemas/RoleCriterion'
        - type: object
          required: [definitionHash]
          properties:
            definitionHash:
              {
                type: string,
                readOnly: true,
                description: Server-calculated SHA-256 of the complete criterion definition.,
              }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
          enum:
            [
              validation_error,
              not_found,
              unauthenticated,
              forbidden,
              csrf_failed,
              rate_limited,
              conflict,
              unavailable,
              internal_error,
            ]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    SearchBriefView:
      type: object
      required: [profileVersion, rubricVersion, rubricSource, requirements, facts]
      properties:
        profileVersion: { type: integer, format: int64 }
        rubricVersion: { type: string }
        rubricSource: { type: string }
        catalogVersion: { type: string }
        requirements:
          type: array
          items: { $ref: '#/components/schemas/RoleCriterionView' }
        facts:
          type: array
          items:
            type: object
            required: [key, value]
            properties:
              key: { type: string }
              value: { type: string }
    ReasonChoice:
      type: object
      required: [id, label, detail]
      properties:
        id: { type: string }
        label: { type: string }
        detail: { type: string }
    ReasonCatalogView:
      type: object
      required:
        [profileVersion, rubricVersion, catalogVersion, positive, negative, missingInformation]
      properties:
        profileVersion: { type: integer, format: int64 }
        rubricVersion: { type: string }
        catalogVersion: { type: string }
        positive: { type: array, items: { $ref: '#/components/schemas/ReasonChoice' } }
        negative: { type: array, items: { $ref: '#/components/schemas/ReasonChoice' } }
        missingInformation: { type: array, items: { $ref: '#/components/schemas/ReasonChoice' } }
    FindingReason:
      type: object
      required: [reasonId, kind, label, detail, jevSupport]
      properties:
        reasonId: { type: string }
        kind: { type: string, enum: [positive, negative, missing_information] }
        label: { type: string }
        detail: { type: string }
        jevSupport: { type: number }
    FindingEvidenceLink:
      type: object
      required: [captureId, spanStart, spanEnd, excerptSha256]
      properties:
        captureId: { type: string }
        spanStart: { type: integer }
        spanEnd: { type: integer }
        excerptSha256: { type: string }
    FindingEntry:
      type: object
      required:
        [
          opportunityId,
          opportunityRevision,
          group,
          assessmentId,
          profileVersion,
          rubricVersion,
          catalogVersion,
          reasons,
          evidenceLinks,
          stale,
        ]
      properties:
        opportunityId: { type: string }
        opportunityRevision: { type: integer, format: int64 }
        group:
          type: string
          enum:
            [recommended, could_be_recommended, probably_not_recommended, not_recommended, unknown]
        unknownBasis: { type: string }
        assessmentId: { type: string }
        profileVersion: { type: integer, format: int64 }
        rubricVersion: { type: string }
        catalogVersion: { type: string }
        reasons: { type: array, items: { $ref: '#/components/schemas/FindingReason' } }
        conflict: { $ref: '#/components/schemas/ReasonChoice' }
        missingFact: { $ref: '#/components/schemas/ReasonChoice' }
        evidenceLinks:
          type: array
          items: { $ref: '#/components/schemas/FindingEvidenceLink' }
        sourceRef:
          type: object
          required: [sourceId, sourceRevision]
          properties:
            sourceId: { type: string }
            sourceRevision: { type: string }
            observedUrl: { type: string }
        stale: { type: boolean }
        staleBasis: { type: string }
    FindingList:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/FindingEntry' } }
        nextCursor: { type: string }
...